refactor: rename widget files and update import paths for consistency
