+++ lib/src/views/home/home_seg2_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'widgets/d_day_widget/d_day_list_tile_module.dart';
+import 'widgets/d_day_widget/d_day_list_tile_widget.dart';
 
 //Dday
 class Seg2Screen extends StatelessWidget {
