+++ lib/src/views/home/home_seg3_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import 'widgets/daily_summary/daily_summary_module.dart';
+import 'widgets/daily_summary/daily_summary_widget.dart';
 
 //요약
 class Seg3Screen extends StatelessWidget {
