+++ lib/src/views/home/widgets/daily_summary/daily_summary_widget.dart
@@ -5,9 +5,9 @@ import '../../../../../core/enum/mxnRate.dart';
 import '../../../../../core/widgets/mxnContainer.dart';
 import '../../../../viewmodels/study_record/study_record_viewmodel.dart';
 import '../../../../viewmodels/study_setting/study_setting_view_model.dart';
-import '../../../statistics/widgets/bar_charts/study_bar_chart_module.dart';
-import '../../../statistics/widgets/pie_charts/study_pie_chart_module.dart';
-import 'goal_progress_module.dart';
+import '../../../statistics/widgets/bar_charts/study_bar_chart_widget.dart';
+import '../../../statistics/widgets/pie_charts/study_pie_chart_widget.dart';
+import 'goal_progress_widget.dart';
 
 class DailySummaryWidget extends ConsumerWidget {
   const DailySummaryWidget({super.key});
