+++ lib/src/views/home/widgets/daily_summary/goal_progress_widget.dart
