+++ lib/src/views/more/widgets/settings/study_setting_widget.dart
@@ -4,10 +4,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../../core/enum/mxnRate.dart';
-import '../../../../../viewmodels/study_setting/study_setting_view_model.dart';
-import '../setting_container_widget.dart';
-import '../setting_tile_widget.dart';
+import '../../../../../core/enum/mxnRate.dart';
+import '../../../../viewmodels/study_setting/study_setting_view_model.dart';
+import 'setting_container_widget.dart';
+import 'setting_tile_widget.dart';
 
 class StudySettingWidget extends ConsumerWidget {
   const StudySettingWidget({super.key});
