+++ lib/src/views/statistics/widgets/bar_charts/study_bar_chart_widget.dart
